+++ tests/testthat/test-github-actions.R
@@ -124,3 +124,22 @@ test_that("use_tidy_github_actions() configures the full check and pr commands",
   expect_proj_file(".github/workflows/pr-commands.yaml")
   expect_proj_file(".github/workflows/pkgdown.yaml")
 })
+
+
+test_that("use_github_action() allows for custom urls", {
+  skip_if_no_git_config()
+
+  scoped_temporary_package()
+  use_git()
+  use_git_remote(name = "origin", url = "https://github.com/fake/fake")
+  use_description_field("URL", "https://github.com/fake/fake")
+  use_readme_md()
+
+  # Directly call to r-lib actions
+  use_github_action("check-full.yaml",
+                    "https://raw.githubusercontent.com/r-lib/actions/master/examples")
+  expect_proj_dir(".github")
+  expect_proj_dir(".github/workflows")
+  expect_proj_file(".github/workflows/check-full.yaml")
+
+})